Remove placeholder implementations for certain object
Many polyfills check an object's prototype to see if a certain function is
defined before applying themselves. In this case, since NJS defines the functions
on the object prototypes with implementations that crash, polyfills may not be applied
as-is during the bundling/transpilation step if including code not supported by the NJS
runtime natively.